woops, fixed batt stuff
git-svn-id: https://conky.svn.sourceforge.net/svnroot/conky/trunk/conky1@623 7f574dfc-610e-0410-a909-a81674777703

+++ src/linux.c
@@ -1264,7 +1264,7 @@ static char pb_battery_info[3][32];
 static double pb_battery_info_update;
  
 #define PMU_PATH "/proc/pmu"
-void get_powerbook_batt_info(char *buf, unsigned int n, int i)
+void get_powerbook_batt_info(char *buf, size_t n, int i)
 {
         static int rep;
         const char* batt_path = PMU_PATH "/battery_0";
